pyproject.toml
[project]
name = "ursa-ai"
dynamic = ["version"]
description = "Agents for science at LANL"
readme = "README.md"
authors = [
    { name = "Mike Grosskopf", email = "mikegros@lanl.gov" },
    { name = "Nathan Debardeleben", email = "ndebard@lanl.gov" },
    { name = "Rahul Somasundaram", email = "rsomasundaram@lanl.gov" },
    { name = "Isaac Michaud", email = "imichaud@lanl.gov" },
    { name = "Avanish Mishra", email = "avanish@lanl.gov" },
    { name = "Arthur Lui", email = "alui@lanl.gov" },
    { name = "Russell Bent", email = "rbent@lanl.gov" },
    { name = "Earl Lawrence", email = "earl@lanl.gov" }
]
license = "BSD-3-Clause"
license-files = [ "LICENSE" ]
requires-python = ">=3.10"
dependencies = [
    "arxiv>=2.2.0,<3.0",
    "beautifulsoup4>=4.13.4,<5.0",
    "randomname>=0.2.1,<0.3",
    "pandas>=2.3.1,<3.0",
    "pillow>=11.3.0,<12.0",
    "pymupdf>=1.26.0,<2.0",
    "pypdf>=5.9.0,<6.0",
    "rich>=13.9.4,<14.0",
    "mp-api>=0.45.8,<0.46",
    "ddgs>=9.5.5",
    "typer>=0.16.1",
    "trafilatura>=1.6.1,<1.7",
    "selectolax>=0.4.0,<0.5",
    # "langchain-google-genai>=2.1.9,<3.0",
    # "langchain-anthropic>=0.3.19,<0.4",
    "langchain>=1.0.3",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.1",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "langchain-text-splitters>=1.0.0",
    "langchain-ollama>=1.0.0",
    "langchain-mcp-adapters>=0.1.12,<0.2",
    "pydantic>=2.12.0,<3.0",
    "pydantic-settings>=2.11.0",
    "mcp>=1.20.0,<2.0",
    "aiosqlite>=0.21.0",
    "asyncstdlib>=3.13.1",
    "pyyaml>=6.0.3",
]
classifiers = [
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]

[project.scripts]
ursa = "ursa.cli:main"

[project.urls]
Homepage = "https://github.com/lanl/ursa"
Documentation = "https://github.com/lanl/ursa/tree/main/docs"
Repository = "https://github.com/lanl/ursa"
Issues = "https://github.com/lanl/ursa/issues"

[project.optional-dependencies]
fm = ["torch>=2.9.0"] # Ursa's MCP FM interface

[build-system]
requires = ["setuptools>=74.1,<80", "setuptools-git-versioning>=2.0,<3"]
build-backend = "setuptools.build_meta"

[tool.setuptools-git-versioning]
enabled = true

[tool.ruff]
line-length = 80
preview = true

[tool.ruff.lint]
extend-select = ["I", "TID", "UP006", "UP007"]  # W505 (doc-line-too-long); D (pydocstyle)
pydocstyle.convention = "numpy"
pycodestyle.max-doc-length = 80

[tool.ruff.lint.flake8-tidy-imports.banned-api]
"typing_extensions" = { msg = "Use standard typing features instead of `typing_extensions`." }

# Ignore test file documentation linting.
[tool.ruff.lint.extend-per-file-ignores]
"tests/**/*.py" = ["D"]

[tool.setuptools.package-data]
ursa = ["observability/pricing.json"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "fastmcp>=2.13.3",
    "notebook>=7.3.3",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.3.0",
    "scikit-optimize>=0.10.2",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-autorefs>=1.4.3",
    "mkdocs-material>=9.6.21",
    "mkdocstrings-python>=1.18.2",
]
lammps = [
    "atomman>=1.5.2",
]
opt = [
    "ortools>=9.14,<9.15",
]
